+++ TEAM_1/hson_prac/Makefile
@@ -30,8 +30,13 @@ INCLUDE = -I./ -I./libft
 			bmp.c
 #SRCS = $(addprefix $(SRCS_DIR), $(SRCS_NAME))
 
-SRCS = cd.c \
-		error.c
+SRCS = exec_cd.c \
+		exec_echo.c \
+		exec_env.c \
+		exec_exit.c \
+		exec_pwd.c \
+		error.c \
+		main.c \
 
 #OBJS_DIR = ./obj/
 #OBJS_NAME = $(SRCS_NAME:.c=.o)
