+++ TEAM_1/hson_prac/exec_export.c
@@ -1,0 +1,3 @@
+#include "minishell.h"
+
+int		exec_export(t_cmd *cmd, t_env *env);
